+++ sn/des5y.py
@@ -31,7 +31,7 @@ def Ez(params):
 def DM_z(z, params):
     dh_grid = (c / H0) / Ez(params)
     dy = (dh_grid[:-1] + dh_grid[1:]) / 2
-    cum_dm = np.zeros(z_grid.size)
+    cum_dm = np.zeros(z_grid.size, dtype=np.float64)
     cum_dm[1:] = np.cumsum(dx * dy)
     return interp_hermite(z, z_grid, cum_dm, dh_grid)
