+++ sn/pantheon_and_sh0es.py
@@ -32,7 +32,7 @@ def Ez(params):
 def DM_z(theta):
     dh_grid = (c / theta[1]) / Ez(theta)
     dy = (dh_grid[:-1] + dh_grid[1:]) / 2
-    cum_dm = np.zeros(z_grid.size)
+    cum_dm = np.zeros(z_grid.size, dtype=np.float64)
     cum_dm[1:] = np.cumsum(dx * dy)
     return interp_hermite(z_values, z_grid, cum_dm, dh_grid)
